+++ test/controllers/api/v1/nets_controller_test.rb
@@ -36,7 +36,7 @@ module Api
       end
 
       test "should respond with error object when Content-Type is wrong" do
-        error_object = Api::V1::Exceptions::WrongContentTypeError.new
+        error_object = Api::V1::Exceptions::InvalidContentTypeError.new
         response_json = RequestErrorSerializer.new([error_object], message: error_object.title).serialized_json
 
         get api_v1_nets_url, headers: { "Content-Type": "text/plain" }
@@ -51,7 +51,7 @@ module Api
       end
 
       test "should respond with error object when Accept is wrong" do
-        error_object = Api::V1::Exceptions::WrongAcceptError.new
+        error_object = Api::V1::Exceptions::InvalidAcceptError.new
         response_json = RequestErrorSerializer.new([error_object], message: error_object.title).serialized_json
 
         get api_v1_nets_url, headers: { "Content-Type": "application/vnd.api+json", "Accept": "application/json" }